# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/cars": {
            "post": {
                "description": "Post details of car corresponding to the input Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Post details for given Id",
                "parameters": [
                    {
                        "description": "create car",
                        "name": "models.Car",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    }
                }
            }
        },
        "/cars/{Id}": {
            "get": {
                "description": "Get details for car corresponding to the input Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Get details for a given Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the car",
                        "name": "Id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    }
                }
            }
        },
        "/cars/{id}": {
            "delete": {
                "description": "Delete the order corresponding to the input Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Delete car identified by given id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the car to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "description": "Update the car corresponding to the input Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Update car identified by the given Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the car to be updated",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "description": "Get details of all cars",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Get details",
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Car": {
            "type": "object",
            "properties": {
                "harga": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "merk": {
                    "type": "string"
                },
                "pemilik": {
                    "type": "string"
                },
                "typecars": {
                    "type": "string"
                }
            }
        }
    }
}